refactor(en-scribe): extract HTTP server factory (mirror of en-quire)
Duplicate of en-quire/src/http-server.ts on the en-scribe side, rather
than refactoring to a shared en-core module. The handlers are 90%+
identical, but the bins are published as separate packages and a
cross-package shared implementation would couple all three. Explicit
duplication is the trade-off: keep the copies in sync by hand;
divergence is a signal to think twice rather than drift quietly.

Changes:

- packages/en-scribe/src/http-server.ts — new file, parallel to the
  en-quire version (PR #72). Exports `createMcpHttpServer` taking the
  same options shape; `realm` argument ensures WWW-Authenticate says
  "en-scribe".

- packages/en-scribe/src/bin.ts — startHttpTransport is now a thin
  wrapper that calls the factory, binds listen(), and wires SIGTERM
  shutdown. Dropped direct imports of node:http / MCP SDK transport /
  node:crypto / authenticateBearer — all now handled inside the
  factory.

No behaviour change. 509/509 suite green, lint clean.

Follow-up (not in this PR): integration tests mirroring en-quire's
(PR #72) to cover en-scribe's auth wiring end-to-end.
